# WebContent/restsql-swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.9.2",
    "title": "easyREST Web Service",
    "description": "easyrest is an source, ultra-lightweight data access layer for HTTP clients",
    "contact": "contato@objetivait.com.br",
    "license": {
      "name": "MIT",
      "url": ""
    }
  },
  "host": null,
  "basePath": "\/easyrest",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application\/json",
    "application\/xml"
  ],
  "produces": [
    "application\/json",
    "application\/xml"
  ],
...
  "paths": {
    "\/res\/{resName}": {
      "get": {
        "tags": [
          "data"
        ],
        "description": "Returns rows for Resource matching query parameters. See [ResGetQuery API doc] for examples.\n* Example: \/res\/sdk.film.Film?year=2010\n* Example: \/res\/sdk.film.Film?_limit=100&_offset=0\n\nNote: Each column name, or configured column alias, specified in the resource definition is available for query, however you will not be able to enter those in this Swagger interface.",
        "operationId": "ResGetQuery",
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "name": "_limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32",
            "description": "Limits number of records in response. Must be paired with _offset. See [ResGetQuery API doc]for detail.",
            "default": 10
          },
          {
            "name": "_offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32",
            "description": "Sets the start row in response, where the fit row is zero. Must be paired with _limit. See [ResGetQuery API doc] for detail.",
            "default": 10
          },
          {
            "$ref": "#\/parameters\/outputParam"
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response",
            "schema": {
              "$ref": "#\/definitions\/ReadResponse"
            }
          },
          "400": {
            "description": "Invalid parameters",
            "schema": {
              
            }
          },
          "404": {
            "description": "Invalid  Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": " Exception \/ database error",
            "schema": {
              
            }
          }
        }
      },
      "post": {
        "tags": [
          "data"
        ],
        "description": "Creates row using URL-encoded form params or one or more rows in an XML or JSON body. See [ResPostUrlEnc API doc]* Example: POST \/res\/sdk.film.Film\/, XML request body - &lt;request&gt;&lt;film title=\"The Fox\" year=\"2018\" language_id=\"1\"\/&gt;&lt;\/request&gt;",
        "operationId": "ResPostQuery",
        "consumes": [
          "application\/json",
          "application\/xml",
          "application\/x-www-form-urlencoded"
        ],
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Array of Resource-specific objects"
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response includes affected rows and data created as an array of Resource-specific objects",
            "schema": {
              "$ref": "#\/definitions\/WriteResponseExt"
            }
          },
          "404": {
            "description": "Invalid Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": "Exception \/ database error",
            "schema": {
              
            }
          }
        }
      },
      "put": {
        "tags": [
          "data"
        ],
        "description": "Updates rows using query params and URL-encoded form params. See [ResPutQueryUrlEnc API doc]* Example: PUT \/res\/sdk.film.Film\/, XML request body - &lt;request&gt;&lt;film id=\"1001\" title=\"Grey Fox\" year=\"2017\"\/&gt;&lt;\/request&gt;",
        "operationId": "ResPutQuery",
        "consumes": [
          "application\/json",
          "application\/xml",
          "application\/x-www-form-urlencoded"
        ],
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Array of Resource-specific objects"
            }
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response",
            "schema": {
              "$ref": "#\/definitions\/WriteResponse"
            }
          },
          "400": {
            "description": "Invalid parameters",
            "schema": {
              
            }
          },
          "404": {
            "description": "Invalid Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": "Exception \/ database error",
            "schema": {
              
            }
          }
        }
      },
      "delete": {
        "tags": [
          "data"
        ],
        "description": "Deletes one or more rows using query params. See [ResDeleteQuery API doc] for detail.\n* Example: DELETE \/res\/sdk.film.Film?title=Red%20Fox.\n\nNote: Each column name, or configured column alias, specified in the resource definition is available for query, however you will not be able to enter those in this Swagger interface.\n\nOr deletes all rows in an XML or JSON body. See [ResDeleteXmlOrJson API doc] for detail.\n* Example: DELETE \/res\/sdk.film.Film, JSON request body - {\"films\": [ { \"id\": 1001 }, { \"id\": 1002 } ] }\n* Example: DELETE \/res\/sdk.film.Film\/, XML request body - &lt;request&gt;&lt;film id=\"1001\"\/&gt;&lt;film id=\"1002\"\/&gt;&lt;\/request&gt;",
        "operationId": "ResDeleteQuery",
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Array of Resource-specific objects"
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response",
            "schema": {
              "$ref": "#\/definitions\/WriteResponse"
            }
          },
          "404": {
            "description": "Invalid Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": "Exception \/ database error",
            "schema": {
              
            }
          }
        }
      }
    },
    "\/res\/{resName}\/{resId}": {
      "get": {
        "tags": [
          "data"
        ],
        "description": "Returns row for Resource matching primary key. See [ResGetPath API doc] for detail.\n* Example: \/res\/sdk.film.Film\/1001",
        "operationId": "ResGetPath",
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "$ref": "#\/parameters\/resIdParam"
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response",
            "schema": {
              "$ref": "#\/definitions\/ReadResponse"
            }
          },
          "404": {
            "description": "Invalid Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": "Exception \/ database error",
            "schema": {
              
            }
          }
        }
      },
      "post": {
        "tags": [
          "data"
        ],
        "description": "Creates children rows using XML or JSON body for hierarchical SQL Resources. See [ResPostPathXmlOrJson API doc] for detail.",
        "operationId": "ResPostPath",
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Array of Resource-specific objects"
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          },
          {
            "$ref": "#\/parameters\/resIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response includes affected rows and data created as an array of Resource-specific objects",
            "schema": {
              "$ref": "#\/definitions\/WriteResponseExt"
            }
          },
          "404": {
            "description": "Invalid Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": "Exception \/ database error",
            "schema": {
              
            }
          }
        }
      },
      "put": {
        "tags": [
          "data"
        ],
        "description": "Updates specific row using URL-encoded form params. See [ResPutPathUrlEnc API doc] for detail.",
        "operationId": "ResPutPath",
        "consumes": [
          "application\/json",
          "application\/xml",
          "application\/x-www-form-urlencoded"
        ],
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Array of Resource-specific objects"
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          },
          {
            "$ref": "#\/parameters\/resIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response",
            "schema": {
              "$ref": "#\/definitions\/WriteResponse"
            }
          },
          "400": {
            "description": "Invalid parameters",
            "schema": {
              
            }
          },
          "404": {
            "description": "Invalid Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": "Exception \/ database error",
            "schema": {
              
            }
          }
        }
      },
      "delete": {
        "tags": [
          "data"
        ],
        "description": "Deletes specific row by primary key. See [ResDeletePath API doc] for detail.\n* Example: DELETE \/res\/sdk.film.Film\/1001\n\nFor hierarchical Resources, deletes all children contained in the XML or JSON body. See [ResDeletePathXmlOrJson API doc] for detail.",
        "operationId": "ResByResNameAndResIdDelete",
        "produces": [
          "application\/json",
          "application\/xml"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Array of Resource-specific objects"
          },
          {
            "$ref": "#\/parameters\/resNameParam"
          },
          {
            "$ref": "#\/parameters\/resIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Normal response",
            "schema": {
              "$ref": "#\/definitions\/WriteResponse"
            }
          },
          "404": {
            "description": "Invalid Resource name",
            "schema": {
              
            }
          },
          "500": {
            "description": "Exception \/ database error",
            "schema": {
              
            }
          }
        }
      }
    },
...
  },
...
}